+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Major Changes
 
 ### Minor Changes
+- Added the `NftNamespace.getTransfersForOwner()` method to get all NFT transfers to or from a provided owner address.
+- Deprecated the `GetMintedNftsResponse` interface in favor of the `TransfersNftResponse`. The `TransfersNftResponse` contains the same properties as the `GetMintedNftsResponse` and includes additional fields about the transfer.
 
 ## 2.4.0